v. 1.11

+++ src/main/java/com/epam/car_rental/dao/car/CarDAO.java
@@ -18,5 +18,5 @@ public interface CarDAO {
 
     List<Car> getCarByType(Car.Type type) throws DAOException;
 
-    long getCarIdByModel(String model) throws DAOException;
+    int getCarIdByModel(String model) throws DAOException;
 }
